Fixed issue with unordered values in matrix_list

+++ matrix_utils.h
@@ -7,21 +7,21 @@
 
 namespace matrix
 {
-    void print_matrix(std::vector<std::vector<int>> matrix);
+    void print_matrix(std::vector<std::vector<double>> matrix);
 
-    std::vector<std::vector<int>> transpose(std::vector<std::vector<int>> matrix);
+    std::vector<std::vector<double>> transpose(std::vector<std::vector<double>> matrix);
 
-    std::vector<std::vector<int>> add(std::vector<std::vector<int>> matrix_A, std::vector<std::vector<int>> matrix_B);
+    std::vector<std::vector<double>> add(std::vector<std::vector<double>> matrix_A, std::vector<std::vector<double>> matrix_B);
 
-    std::vector<std::vector<int>> subtract(std::vector<std::vector<int>> matrix_A, std::vector<std::vector<int>> matrix_B);
+    std::vector<std::vector<double>> subtract(std::vector<std::vector<double>> matrix_A, std::vector<std::vector<double>> matrix_B);
 
-    std::vector<std::vector<int>> multiply(std::vector<std::vector<int>> matrix_A, std::vector<std::vector<int>> matrix_B);
+    std::vector<std::vector<double>> multiply(std::vector<std::vector<double>> matrix_A, std::vector<std::vector<double>> matrix_B);
 
-    std::vector<std::vector<int>> horizontal_flip(std::vector<std::vector<int>> matrix);
+    std::vector<std::vector<double>> horizontal_flip(std::vector<std::vector<double>> matrix);
 
-    std::vector<std::vector<int>> vertical_flip(std::vector<std::vector<int>> matrix);
+    std::vector<std::vector<double>> vertical_flip(std::vector<std::vector<double>> matrix);
 
-    int determinant(std::vector<std::vector<int>> matrix);
+    double determinant(std::vector<std::vector<double>> matrix);
 
 }
